test: rename phase-numbered test files by the domain they cover
test_phase36_cmd.py through test_phase165_cmd.py encoded development
chronology, not behavior; finding a command's tests meant grep. Each of
the 20 files now carries its subsystem name (center readiness, release
trains, repos fleet, security template audit, daily driver, phase
ledger), with the three cross-module imports updated. Contents are
otherwise untouched; collection count is unchanged.
