+++ tests/test_center_actions_cmd.py
