+++ tests/test_center_readiness_cmd.py
