+++ tests/test_center_report_cmd.py
