+++ tests/test_daily_driver_cmd.py
