+++ tests/test_release_smoke_health_cmd.py
