+++ tests/test_roadmap_inventory_cmd.py
