+++ tests/test_security_template_audit_cmd.py
