+++ tests/test_waiver_policy_health_cmd.py
@@ -5,7 +5,7 @@ from brigade import cli
 from brigade import repos_cmd
 from brigade import work_cmd
 
-from tests.test_phase44_cmd import _build_train
+from tests.test_release_train_evidence_cmd import _build_train
 
 
 def _iso_after(days: int) -> str:
